scraper-ai/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.12.0
# Database connectivity - try different approach if binary fails
psycopg2-binary==2.9.9
# Alternative: psycopg[binary]==3.1.14
python-dotenv==1.0.1
feedparser==6.0.11
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
nltk==3.8.1
schedule==1.2.1
newspaper3k==0.2.8
textblob==0.17.1
asyncio==3.4.3
aiohttp==3.9.3
redis==5.0.1
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
